Implemented reproject_into.
